Update ConfigureServiceAccounts.ps1
This is a bug where $VM was used but was not defined by the function. If you were lucky enough to have $VM defined previously, it would still work. 

This is a bug in the SharePoint configure service accounts that doesn't exist in the SQL or AD configure scripts so the fix is only required here.